fix(mod.go): move module to /Proxy
