+++ Proxy/internal/data.go
@@ -1,4 +1,4 @@
-package proxy
+package data
 
 func TestData() string {
 	return "Hello"
